requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
flask-cors
pyotp
qrcode[pil]
requests